Implementation of Off-axis stereo method
The current stereo implementation is known as Toe-In. In this method the cameras are rotated towards a common convergence point. It gives an impression of 3D, but it is not considered correct.

The Off-axis method uses shifted projection matrices for the individual eyes, and the convergence happens at a plane at a given distance ("3d_convergence_distance").

This initial commit affects the pageflip stereo mode only, the other
modes will be adapted soon